+++ lib/square_connect.dart
@@ -1,9 +1,15 @@
+import 'dart:convert';
+
 import 'package:http/http.dart';
 import 'package:http/io_client.dart';
+import 'package:square_connect/src/auth/auth-enums-converter.dart';
+import 'package:square_connect/src/auth/auth-enums.dart';
+import 'package:square_connect/src/auth/auth-return-objects.dart';
 
 import 'package:square_connect/src/catalog/catalog.dart';
 import 'package:square_connect/src/customer/customer.dart';
 import 'package:square_connect/src/employees/employees.dart';
+import 'package:square_connect/src/helper-classes.dart';
 import 'package:square_connect/src/inventory/inventory.dart';
 import 'package:square_connect/src/labor/labor.dart';
 import 'package:square_connect/src/locations/locations.dart';
@@ -52,6 +58,7 @@ export 'package:square_connect/src/inventory/inventory.dart';
 
 export 'package:square_connect/src/shared-enums.dart';
 export 'package:square_connect/src/shared-objects.dart';
+export 'package:square_connect/src/auth/auth-enums.dart';
 
 /// The Square Connect Client object. It should be obtained by `SquareConnect.instance` or `SquareConnect.getInstanceWithTokens(String token)`
 class SquareConnect {
@@ -71,7 +78,8 @@ class SquareConnect {
     this._authToken = token;
   }
 
-  SquareConnect._withRefreshToken(String refreshToken, String clientId, String clientSecret) {
+  SquareConnect._withRefreshToken(
+      String refreshToken, String clientId, String clientSecret) {
     this._client = IOClient();
     this._refreshToken = refreshToken;
     this._clientId = clientId;
@@ -89,8 +97,10 @@ class SquareConnect {
   }
 
   /// Alternate Entry point for API. Creates an API client to be reused and get fresh auth token from refresh token
-  static SquareConnect getInstanceWithRefreshToken(String refreshToken, String clientId, String clientSecret) {
-    return SquareConnect._withRefreshToken(refreshToken, clientId, clientSecret);
+  static SquareConnect getInstanceWithRefreshToken(
+      String refreshToken, String clientId, String clientSecret) {
+    return SquareConnect._withRefreshToken(
+        refreshToken, clientId, clientSecret);
   }
 
   /// Sets the authorization token of the client for use in all calls. Must be set before any other calls are made.
@@ -103,43 +113,100 @@ class SquareConnect {
     this._client = client;
   }
 
+  /// Gets url for oauth permissions and authorizations.
+  String getAuthEndpoint({
+    String clientId,
+    List<OAuthPermission> permissions,
+    String locale,
+    bool session,
+    String state,
+  }) {
+    var queryParams = [
+      if (clientId != null) QueryParam('client_id', clientId),
+      if (permissions != null)
+        QueryParam(
+            'scope',
+            permissions
+                .map((permission) => getStringFromOAuthPermission(permission))
+                .join(' ')),
+      if (locale != null) QueryParam('locale', locale),
+      if (session != null) QueryParam('session', session.toString()),
+      if (state != null) QueryParam('state', state),
+    ];
+
+    var queryString = getParamListString(queryParams);
+
+    return 'https://connect.squareup.com/oauth2/authorize$queryString';
+  }
+
+  Future<CreateMobileAuthorizationCodeResponse> createMobileAuthorizationCode({
+    String locationId,
+  }) async {
+    if (locationId == null) throw ArgumentError('locationId must not be null');
+
+    var body = {'location_id': locationId};
+
+    var obj = RequestObj(
+      token: _authToken,
+      path: '/mobile/authorization-code',
+      method: RequestMethod.post,
+      client: _client,
+      refreshToken: _refreshToken,
+      clientId: _clientId,
+      clientSecret: _clientSecret,
+      body: body,
+    );
+
+    var response = await obj.makeCall();
+    return CreateMobileAuthorizationCodeResponse.fromJson(
+        json.decode(response.body));
+  }
+
   /// Getter for Catalog API methods.
   CatalogApi get catalogApi {
-    return CatalogApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return CatalogApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Customers API methods.
   CustomersApi get customersApi {
-    return CustomersApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return CustomersApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Locations API methods.
   LocationsApi get locationsApi {
-    return LocationsApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return LocationsApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Orders API methods.
   OrdersApi get ordersApi {
-    return OrdersApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return OrdersApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Transactions API methods.
   TransactionsApi get transactionsApi {
-    return TransactionsApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return TransactionsApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Employees API methods.
   EmployeesApi get employeesApi {
-    return EmployeesApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return EmployeesApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Labor API methods.
   LaborApi get laborApi {
-    return LaborApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return LaborApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 
   /// Getter for Inventory API methods.
   InventoryApi get inventoryApi {
-    return InventoryApi(_authToken, _client, _refreshToken, _clientId, _clientSecret);
+    return InventoryApi(
+        _authToken, _client, _refreshToken, _clientId, _clientSecret);
   }
 }