+++ lib/src/helper-classes.dart
@@ -51,22 +51,46 @@ class RequestObj {
     return _baseUrl + path + getParamListString(queryParams);
   }
 
+  get postHeaders {
+    return {
+      'Authorization': 'Bearer ' + this.token,
+      'Content-Type': 'application/json'
+    };
+  }
+
   get headers {
-    return {'Authorization': 'Bearer ' + this.token};
+    return {
+      'Authorization': 'Bearer ' + this.token,
+    };
   }
 
   Future<Response> makeCall() async {
-    if (refreshToken != null) {
-      var resp = await client.post('$_baseUrl/oauth2/token', body: json.encode({
-        'client_id': clientId,
-        'client_secret': clientSecret,
-        'grant_type': 'refresh_token',
-        'refresh_token': refreshToken,
-      }));
+    if (token == null || refreshToken != null) {
+      var resp = await client.post('$_baseUrl/oauth2/token',
+          body: json.encode({
+            'client_id': clientId,
+            'client_secret': clientSecret,
+            'grant_type': 'refresh_token',
+            'refresh_token': refreshToken,
+          }),
+          headers: {'Content-Type': 'application/json'});
       this.token = json.decode(resp.body)['access_token'];
-    }
-
-    switch (this.method) {
+      switch (this.method) {
+        case RequestMethod.get:
+          return client.get(this.url, headers: this.headers);
+        case RequestMethod.post:
+          return client.post(this.url,
+              headers: this.postHeaders, body: json.encode(this.body));
+        case RequestMethod.delete:
+          return client.delete(url, headers: this.headers);
+        case RequestMethod.put:
+          return client.put(this.url,
+              headers: this.postHeaders, body: json.encode(this.body));
+        default:
+          throw ArgumentError('Method is unsuported');
+      }
+    } else {
+      switch (this.method) {
         case RequestMethod.get:
           return client.get(this.url, headers: this.headers);
         case RequestMethod.post:
@@ -80,6 +104,7 @@ class RequestObj {
         default:
           throw ArgumentError('Method is unsuported');
       }
+    }
   }
 }