+++ lib/src/orders/orders-objects.dart
@@ -192,9 +192,8 @@ class OrderLineItem {
     return OrderLineItem(
       uid: json['uid'],
       name: json['name'],
-      quantity: json['quantity'] != null
-        ? double.parse(json['quantity'])
-        : null,
+      quantity:
+          json['quantity'] != null ? double.parse(json['quantity']) : null,
       orderQuantityUnit: json['quantity_unit'] != null
           ? OrderQuantityUnit.fromJson(json['quantity_unit'])
           : null,